docs: Adds a technical decision record for namespacing
The hope is that the namespacing approach used is intuitive but to make it explicit, I've written a architectural decision record to outline the namespacing approach used. This commit also completes namespacing by introducing the `Tfl::Common` module. Closes #46
