+++ app/graphql/types/tfl/entities/journey_planner/path_type.rb
@@ -3,7 +3,7 @@ module Types
     module Entities
       module JourneyPlanner
         class PathType < Types::BaseObject
-          field :elevation, [Types::JpElevationType, null: true], null: true
+          field :elevation, [Types::Tfl::Common::JpElevationType, null: true], null: true
           field :lineString, String, null: true
           field :stopPoints, [Types::Tfl::Entities::IdentifierType, null: true], null: true
         end
